+++ src/hmm/mod.rs
@@ -142,12 +142,11 @@ pub fn main_hmm_learn(opts: HmmLearnOpts) -> Result<(), Box<dyn Error>> {
         sequence_filenames,
     } = opts;
 
-    let actual_sequence_filenames =
-        utl::get_actual_filenames(sequence_filenames, ".seq", "sequences")?;
+    let seq_filenames = utl::resolve_filenames(sequence_filenames, ".seq", "sequences")?;
 
     println!("ECOZ2 C version: {}", version()?);
 
-    println!("num_actual_sequences: {}", actual_sequence_filenames.len());
+    println!("sequences: {}", seq_filenames.len());
     println!("val_auto = {}", val_auto);
 
     set_random_seed(seed);
@@ -159,7 +158,7 @@ pub fn main_hmm_learn(opts: HmmLearnOpts) -> Result<(), Box<dyn Error>> {
     hmm_learn(
         num_states,
         type_,
-        actual_sequence_filenames,
+        seq_filenames,
         epsilon,
         val_auto,
         max_iterations,
@@ -177,25 +176,20 @@ pub fn main_hmm_classify(opts: HmmClassifyOpts) -> Result<(), Box<dyn Error>> {
         sequence_filenames,
     } = opts;
 
-    let actual_model_filenames = utl::get_actual_filenames(model_filenames, ".hmm", "models")?;
+    let hmm_filenames = utl::resolve_filenames(model_filenames, ".hmm", "models")?;
 
-    let actual_sequence_filenames =
-        utl::get_actual_filenames(sequence_filenames, ".seq", "sequences")?;
+    let seq_filenames = utl::resolve_filenames(sequence_filenames, ".seq", "sequences")?;
 
     println!("ECOZ2 C version: {}", version()?);
 
     println!(
         "number of HMM models: {}  number of sequences: {}",
-        actual_model_filenames.len(),
-        actual_sequence_filenames.len()
+        hmm_filenames.len(),
+        seq_filenames.len()
     );
     println!("show_ranked = {}", show_ranked);
 
-    hmm_classify(
-        actual_model_filenames,
-        actual_sequence_filenames,
-        show_ranked,
-    );
+    hmm_classify(hmm_filenames, seq_filenames, show_ranked);
 
     Ok(())
 }